requirements.txt
python-telegram-bot==20.7
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0
